Fixed two bugs:
1. A nitrogen-containing aromatic rings could canonicalize two different
ways when they were canonicalized as part of a molecular fragment.  In
setting up the initial atom graph invarients, implicit H was ignored on
aromatic nitrogens, so the two nitrogens in in c1ncc[nH]1 were not
distinguished and it could canonicalize as either c1ncc[nH]1 or c1[nH]ccn1.
This only occurs when using a bitmap mask to create a fragment SMILES
(otherwise the two nitrogens are distinguished by other bonds).

2. A segfault that happened when the same molecule was canonicalized
multiple times.  A StereoFacade object is created in CanonicalLabels, then the
OBGenericDataType::StereoData that it points to was deleted causing a
segfault the second time a molecule was canonicalized.