Improved the choice of which bonds to designate as hash/wedge bonds around a stereo center.
Previously, it often made a hash/wedge on a bond to a ring or a bond to
another stereo center.  This new algorithm favors bonds to hydrogens, bonds
to terminal heteroatoms, bonds to non-stereo centers, and acyclic bonds.
Hash/wedge bonds to ring atoms and to other stereo centers are avoided as
much as possible.