Remove unescaped quotes in UsePythonTest.cmake
UsePythonTest.cmake contains a template that is used to create a series of cmake files that execute tests for the python language bindings. Quotes within this template need to be escaped, otherwise they fail to appear in the resulting cmake files.

In this context, the unescaped quotes were optional so no error was actually caused. However, as of cmake 2.8.12, the unescaped quotes cause a warning when running cmake: "Argument not separated from preceding token by whitespace."

This change simply removes the unescaped quotes, causing no changes to the resulting cmake files.