Fix use-after-free in xtcformat from Tomas Trnka
XTCFormat::ReadMolecule() used to pass a pointer to the storage of a local
std::vector to OBMol::SetConformers() after the vector itself was destroyed.
This leads to OBMol::_vconf pointing to freed memory, reading garbage
(as mentioned previously in http://forums.openbabel.org/Reading-XTC-format-does-not-seem-to-work-tp3675197p3733137.html ) and finally a crash when the memory is freed for the
second time by OBMol's ~OBMol() or SetConformers().

There's no reason to use a std::vector here when one doesn't want its
automatic lifetime management.