Identify frequencies with header
In the VASP format, frequency parsing was done until tokenization
returned less than two elements.  However, the list of displacements may
be closed with a section such as the following:

	Finite differences POTIM=  2.000000000000000E-002
	LATTYP: Found a simple tetragonal cell.
	ALAT       =    12.7378929000
	C/A-ratio  =     2.5701183985

	Lattice vectors:

that would not break from the frequency-parsing loop and keep parsing
nonexisting data, eventually leading to a crash.

Instead of breaking the loop on 'less than two tokens', we check whether
the header contains one of 'f/i=' or 'f', if it does not, we break
the loop.