Fix bug in the fingerprint algorithm.
If the atom currently being visited in a path through the atoms can
both close a ring back to the first atom, and can also proceed
to a different atom, and the bond enumerator returns the closing
bond first, then the other path was omitted.  That is because after
the path was marked as closed and the ring added to the fingerprint
bit set, the closure is not cleared.  When the next bond is visited,
the path is not added because it still appears to be closed.

In order for the bit to be missed altogether, this problem must
occur in both directions along the path, so the situation is somewhat
rare.

See bug #206