Fix in OBForceField::WeightedRotorSearch.
The WeightedRotorSearch method of the OBForceField class returned wrong
index to the conformer of the lowest energy.  Although the method worked
correctly, it returned incorrect result. I think this must be somehow
related to a change in indexing conformers which did not propagate fully
to dependent classes. The change is minimal.