Move to using if statement and different conditional
Move to using if statement at request of devs. I also think checking
for if the number of conformers is 0 alongside the code comment of
why it might be zero is a better approach than testing number of
rotatable to cover other unknown cases where only the original
input coordinates remain.